Working on stuff. testing mostly

+++ core/src/com/derelictech/impulse/Constants.java
@@ -1,9 +1,14 @@
 package com.derelictech.impulse;
 
 /**
- * Created by Tim on 2016-12-01.
+ * Project: impulse
+ * Package: com.derelictech.impulse
+ * File:    Constants.java
+ * Author:  voxelv
+ * Creation Date: 2016-12-01
+ * Description:
  */
 public class Constants {
-    public final static float VIEWPORT_WIDTH = 1920;
-    public final static float VIEWPORT_HEIGHT = 1080;
+    public final static int VIEWPORT_WIDTH = 1920;
+    public final static int VIEWPORT_HEIGHT = 1080;
 }
